(#321) refactor: daycell 합치기
