+++ src/domain/Reservation/components/reservation-calendar/DayCell.tsx
@@ -7,10 +7,12 @@ import { useCallback, useEffect, useMemo, useState } from 'react';
 import {
   getReservationsBySchedule,
   getSchedulesByDate,
-  ScheduleItem,
+  type ScheduleItem,
   updateReservationStatus,
 } from '@/domain/Reservation/services/reservation-calendar';
 import { BottomSheet } from '@/shared/components/ui/bottom-sheet';
+import Popover from '@/shared/components/ui/popover';
+import { usePopover } from '@/shared/components/ui/popover/PopoverContext';
 import Tabs from '@/shared/components/ui/tabs';
 
 import type {
@@ -28,28 +30,43 @@ interface DayCellProps {
   isLastRow: boolean;
   reservation: Reservation | null;
   selectedActivityId: number | null;
+  displayMode?: 'popover' | 'bottomsheet'; // 🔥 UI 모드 선택
 }
-// BottomSheet 닫기 버튼 컴포넌트
-const CloseButton = ({ onClose }: { onClose: () => void }) => (
+
+// Popover 닫기 버튼
+const PopoverCloseButton = () => {
+  const { setIsOpen } = usePopover();
+  return (
+    <button type='button' onClick={() => setIsOpen(false)}>
+      <X className='size-15 cursor-pointer font-bold' />
+    </button>
+  );
+};
+
+// BottomSheet 닫기 버튼
+const BottomSheetCloseButton = ({ onClose }: { onClose: () => void }) => (
   <button type='button' onClick={onClose}>
     <X className='size-15 cursor-pointer font-bold' />
   </button>
 );
 
-export default function DayCellBottomSheet({
+export default function DayCell({
   day,
   isCurrentMonth,
   isToday,
   isLastRow,
   reservation,
   selectedActivityId,
+  displayMode = 'popover', // 🔥 기본값은 popover
 }: DayCellProps) {
   const queryClient = useQueryClient();
   const [selectedScheduleId, setSelectedScheduleId] = useState<number | null>(
     null,
   );
 
+  // BottomSheet용 상태
   const [isBottomSheetOpen, setIsBottomSheetOpen] = useState(false);
+
   const [activeTab, setActiveTab] = useState<
     'pending' | 'confirmed' | 'declined'
   >('pending');
@@ -84,8 +101,6 @@ export default function DayCellBottomSheet({
       getSchedulesByDate(selectedActivityId!, day.format('YYYY-MM-DD')),
     enabled: !!selectedActivityId,
   });
-
-  console.log('schedules', schedules);
 
   // 스케줄 데이터가 로드되면 첫 번째 스케줄을 선택
   useEffect(() => {
@@ -131,7 +146,7 @@ export default function DayCellBottomSheet({
         declined: declined || [],
       };
     },
-    enabled: !!selectedScheduleId, // 스케줄 ID가 선택되었을 때만 실행
+    enabled: !!selectedScheduleId,
   });
 
   // 🔽 3. '하나 승인 후 나머지 거절' 비즈니스 로직을 처리하는 전용 뮤테이션
@@ -142,13 +157,11 @@ export default function DayCellBottomSheet({
       reservationsToDecline: ReservationItem[];
     }) => {
       const { reservationId, reservationsToDecline } = variables;
-      // 1. 선택한 예약 승인 처리
       await updateReservationStatus({
         activityId: selectedActivityId!,
         reservationId,
         status: 'confirmed',
       });
-      // 2. 나머지 예약들 거절 처리
       await Promise.all(
         reservationsToDecline.map((r) =>
           updateReservationStatus({
@@ -160,7 +173,6 @@ export default function DayCellBottomSheet({
       );
     },
     onSuccess: () => {
-      // 성공 시 관련된 모든 쿼리를 무효화하여 UI 자동 업데이트
       queryClient.invalidateQueries({ queryKey: ['reservationsBySchedule'] });
       queryClient.invalidateQueries({ queryKey: ['schedules'] });
       queryClient.invalidateQueries({ queryKey: ['reservationDashboard'] });
@@ -184,7 +196,7 @@ export default function DayCellBottomSheet({
     onError: (error) => console.error('거절 실패:', error),
   });
 
-  // 🔽 핸들러 함수들은 뮤테이션을 '호출'하는 역할만 담당합니다.
+  // 🔽 핸들러 함수들
   const handleApprove = useCallback(
     (reservationId: number, scheduleId: number) => {
       if (isApproving) return;
@@ -222,7 +234,6 @@ export default function DayCellBottomSheet({
     return counts ?? { pending: 0, confirmed: 0, declined: 0 };
   }, [schedules]);
 
-  // ✅ 그 다음에 displayItems 선언
   const displayItems = useMemo(() => {
     const items: { status: ReservationStatus; count: number }[] = [];
     if (reservationCounts.pending > 0) {
@@ -237,119 +248,144 @@ export default function DayCellBottomSheet({
     return items;
   }, [reservationCounts]);
 
-  console.log('displayItems', displayItems);
-
   const totalReservations =
     reservationCounts.pending +
     reservationCounts.confirmed +
     reservationCounts.declined;
 
-  return (
-    <BottomSheet.Root
-      open={isBottomSheetOpen}
-      onOpenChange={setIsBottomSheetOpen}
+  // 🔥 공통 셀 UI
+  const cellContent = (
+    <div
+      role='gridcell'
+      aria-label={`${day.format('M월 D일')}`}
+      className={styles.cellClasses}
     >
-      <BottomSheet.Trigger>
-        <div
-          role='gridcell'
-          aria-label={`${day.format('M월 D일')}`}
-          className={styles.cellClasses}
-        >
-          {displayItems.length > 0 && (
-            <div className='absolute top-[10%] left-[60%] size-6 rounded-full bg-red-500' />
-          )}
-          <div className={`${styles.dateClasses} font-size-16`}>
-            {day.format('D')}
-          </div>
-          <div className='mt-1 flex w-full flex-col items-center space-y-1'>
-            {displayItems.map((item, index) => (
-              <div
-                key={`${reservation?.date}-${item.status}-${index}`}
-                className={`font-size-14 w-[90%] truncate rounded-xl px-1 text-center font-medium ${getColorClassByStatus(item.status)}`}
-              >
-                {STATUS_LABELS[item.status]} {item.count}명
-              </div>
-            ))}
-          </div>
-        </div>
-      </BottomSheet.Trigger>
-      <BottomSheet.Content>
-        <div className='h-full min-w-[40rem] space-y-3 p-10'>
-          <div className='flex items-center justify-between'>
-            <div className='flex items-end gap-5'>
-              <h3 className='font-size-20 font-bold text-gray-900'>
-                {day.format('YY년 M월 D일')}
-              </h3>
-              <span className='font-size-12 text-gray-500'>
-                {totalReservations}개의 예약
-              </span>
-            </div>
-            <CloseButton onClose={() => setIsBottomSheetOpen(false)} />
-          </div>
-
-          <Tabs.Root
-            defaultValue={activeTab}
-            value={activeTab}
-            onValueChange={(value) =>
-              setActiveTab(value as 'pending' | 'confirmed' | 'declined')
-            }
+      {displayItems.length > 0 && (
+        <div className='absolute top-[10%] left-[60%] size-6 rounded-full bg-red-500' />
+      )}
+      <div className={`${styles.dateClasses} font-size-16`}>
+        {day.format('D')}
+      </div>
+      <div className='mt-1 flex w-full flex-col items-center space-y-1'>
+        {displayItems.map((item, index) => (
+          <div
+            key={`${reservation?.date}-${item.status}-${index}`}
+            className={`font-size-14 w-[90%] truncate rounded-xl px-1 text-center font-medium ${getColorClassByStatus(item.status)}`}
           >
-            <Tabs.List className='font-size-14 flex'>
-              <Tabs.Trigger value='pending'>
-                신청 {reservationCounts.pending}
-              </Tabs.Trigger>
-              <Tabs.Trigger value='confirmed'>
-                승인 {reservationCounts.confirmed}
-              </Tabs.Trigger>
-              <Tabs.Trigger value='declined'>
-                거절 {reservationCounts.declined}
-              </Tabs.Trigger>
-            </Tabs.List>
+            {STATUS_LABELS[item.status]} {item.count}명
+          </div>
+        ))}
+      </div>
+    </div>
+  );
 
-            <Tabs.Content value='pending'>
-              <ReservationDetail
-                schedules={schedules || []}
-                reservations={reservationsByStatus.pending}
-                emptyMessage='신청된 예약이 없습니다.'
-                showApprovalButton={true}
-                showRejectButton={true}
-                onApprove={handleApprove}
-                onReject={handleReject}
-                onTimeSlotSelect={handleTimeSlotSelect}
-                isLoading={isApproving || isRejecting}
-              />
-            </Tabs.Content>
-
-            <Tabs.Content value='confirmed'>
-              <ReservationDetail
-                schedules={schedules || []}
-                reservations={reservationsByStatus.confirmed}
-                emptyMessage='승인된 예약이 없습니다.'
-                showApprovalButton={false}
-                showRejectButton={true}
-                onApprove={handleApprove}
-                onReject={handleReject}
-                onTimeSlotSelect={handleTimeSlotSelect}
-                isLoading={isApproving || isRejecting}
-              />
-            </Tabs.Content>
-
-            <Tabs.Content value='declined'>
-              <ReservationDetail
-                schedules={schedules || []}
-                reservations={reservationsByStatus.declined}
-                emptyMessage='거절된 예약이 없습니다.'
-                showApprovalButton={false}
-                showRejectButton={false}
-                onApprove={handleApprove}
-                onReject={handleReject}
-                onTimeSlotSelect={handleTimeSlotSelect}
-                isLoading={isApproving || isRejecting}
-              />
-            </Tabs.Content>
-          </Tabs.Root>
+  // 🔥 공통 콘텐츠 UI
+  const contentUI = (
+    <div className='tablet:min-w-[40rem] h-full space-y-3 p-10'>
+      <div className='flex items-center justify-between'>
+        <div className='flex items-end gap-5'>
+          <h3 className='font-size-20 font-bold text-gray-900'>
+            {day.format('YY년 M월 D일')}
+          </h3>
+          <span className='font-size-12 text-gray-500'>
+            {totalReservations}개의 예약
+          </span>
         </div>
-      </BottomSheet.Content>
-    </BottomSheet.Root>
+        {displayMode === 'popover' ? (
+          <PopoverCloseButton />
+        ) : (
+          <BottomSheetCloseButton onClose={() => setIsBottomSheetOpen(false)} />
+        )}
+      </div>
+
+      <Tabs.Root
+        defaultValue={activeTab}
+        value={activeTab}
+        onValueChange={(value) =>
+          setActiveTab(value as 'pending' | 'confirmed' | 'declined')
+        }
+      >
+        <Tabs.List className='font-size-14 flex'>
+          <Tabs.Trigger value='pending'>
+            신청 {reservationCounts.pending}
+          </Tabs.Trigger>
+          <Tabs.Trigger value='confirmed'>
+            승인 {reservationCounts.confirmed}
+          </Tabs.Trigger>
+          <Tabs.Trigger value='declined'>
+            거절 {reservationCounts.declined}
+          </Tabs.Trigger>
+        </Tabs.List>
+
+        <Tabs.Content value='pending'>
+          <ReservationDetail
+            schedules={schedules || []}
+            reservations={reservationsByStatus.pending}
+            emptyMessage='신청된 예약이 없습니다.'
+            showApprovalButton={true}
+            showRejectButton={true}
+            onApprove={handleApprove}
+            onReject={handleReject}
+            onTimeSlotSelect={handleTimeSlotSelect}
+            isLoading={isApproving || isRejecting}
+          />
+        </Tabs.Content>
+
+        <Tabs.Content value='confirmed'>
+          <ReservationDetail
+            schedules={schedules || []}
+            reservations={reservationsByStatus.confirmed}
+            emptyMessage='승인된 예약이 없습니다.'
+            showApprovalButton={false}
+            showRejectButton={true}
+            onApprove={handleApprove}
+            onReject={handleReject}
+            onTimeSlotSelect={handleTimeSlotSelect}
+            isLoading={isApproving || isRejecting}
+          />
+        </Tabs.Content>
+
+        <Tabs.Content value='declined'>
+          <ReservationDetail
+            schedules={schedules || []}
+            reservations={reservationsByStatus.declined}
+            emptyMessage='거절된 예약이 없습니다.'
+            showApprovalButton={false}
+            showRejectButton={false}
+            onApprove={handleApprove}
+            onReject={handleReject}
+            onTimeSlotSelect={handleTimeSlotSelect}
+            isLoading={isApproving || isRejecting}
+          />
+        </Tabs.Content>
+      </Tabs.Root>
+    </div>
+  );
+
+  // 🔥 displayMode에 따라 다른 UI 렌더링
+  if (displayMode === 'bottomsheet') {
+    return (
+      <BottomSheet.Root
+        open={isBottomSheetOpen}
+        onOpenChange={setIsBottomSheetOpen}
+      >
+        <BottomSheet.Trigger>{cellContent}</BottomSheet.Trigger>
+        <BottomSheet.Content>{contentUI}</BottomSheet.Content>
+      </BottomSheet.Root>
+    );
+  }
+
+  // 기본값: Popover
+  return (
+    <Popover.Root>
+      <Popover.Trigger>{cellContent}</Popover.Trigger>
+      <Popover.Content
+        position='left-center'
+        withBackdrop
+        className='min-h-[40rem]'
+      >
+        {contentUI}
+      </Popover.Content>
+    </Popover.Root>
   );
 }
